Refactor json generation
Now a single json file per text is generated. These files can be combined
in different ways to get different datasets.
Also, the script does not need to regenerate everything when it is
stopped for some reason.

+++ embem/kafnaftag/naf2json.py
@@ -20,7 +20,7 @@ def create_event(emotion_label, text_id, year):
     group_score = 100
     event_object = {
         'actors': {},
-        'event': event_name(emotion_label, year),
+        'event': event_name(emotion_label, text_id),
         'group': emotion_label,
         'groupName': emotion_label,
         'groupScore': str(group_score),
@@ -89,7 +89,7 @@ def process_emotions(soup, text_id, year, source, em_labels, confidence=0.5):
             if el['resource'] == 'heem' and \
                el['reference'].split(':')[1] in em_labels and \
                c >= confidence:
-                label = event_name(el['reference'].split(':')[1], year)
+                label = event_name(el['reference'].split(':')[1], text_id)
 
                 if label not in events.keys():
                     #print 'created new event', label
@@ -121,7 +121,7 @@ def process_emotions(soup, text_id, year, source, em_labels, confidence=0.5):
 
                 #print ems
                 for e in ems:
-                    label = event_name(e, year)
+                    label = event_name(e, text_id)
                     if label not in events.keys():
                         #print 'created new event2', label
                         events[label] = create_event(e, text_id, year)
@@ -136,7 +136,7 @@ def process_emotions(soup, text_id, year, source, em_labels, confidence=0.5):
 
     print 'found {} events and {} mentions'.format(len(mention_counter.keys()), sum(mention_counter.values()))
     print 'top three events: {}'.format(' '.join(['{} ({})'.format(k, v) for k, v in mention_counter.most_common(3)]))
-    return events, mention_counter
+    return events
 
 
 def get_num_sentences(soup):
@@ -153,25 +153,13 @@ def add_source_text(soup, text_id, json_object):
 
 
 def add_events(events, num_sentences, json_object):
-    new_events = []
-    for event in json_object['timeline']['events']:
-        event_id = event['event']
-        if event_id in events.keys():
-            add_climax_score(events[event_id], num_sentences)
-            new_events.append(merge_events(event, events[event_id]))
-            del events[event_id]
-        else:
-            new_events.append(event)
-
     for event, data in events.iteritems():
         add_climax_score(data, num_sentences)
 
         # TODO: make more intelligent choice for prefLabel (if possible)
         # Also, it seems that prefLabel is not used in the visualization
         data['prefLabel'] = [data['event']]
-        new_events.append(data)
-
-    json_object['timeline']['events'] = new_events
+        json_object['timeline']['events'].append(data)
 
 
 def merge_events(event1, event2):
@@ -195,11 +183,9 @@ def add_climax_score(event, num_sentences):
 @click.command()
 @click.argument('input_dir', type=click.Path(exists=True))
 @click.argument('metadata', type=click.Path(exists=True))
-@click.argument('output_file', type=click.Path())
+@click.argument('output_dir', type=click.Path())
 @click.option('--confidence', default=0.5, help='Confidence value threshold.')
-def run(input_dir, metadata, output_file, confidence):
-    output_dir = os.path.dirname(click.format_filename(output_file))
-
+def run(input_dir, metadata, output_dir, confidence):
     if not os.path.exists(output_dir):
         os.makedirs(output_dir)
 
@@ -218,15 +204,6 @@ def run(input_dir, metadata, output_file, confidence):
     # the labels for which groups are created (emotion labels only)
     emotion_labels = [heem_labels_en[l].lower() for l in heem_emotion_labels]
 
-    json_out = {
-        'timeline': {
-            'events': [],
-            'sources': []
-        }
-    }
-
-    num_events = 0
-
     for i, fi in enumerate(xml_files):
         start = time.time()
 
@@ -234,49 +211,49 @@ def run(input_dir, metadata, output_file, confidence):
         text_id = os.path.basename(fi).split('.')[0]
         text_id = text_id.split('_')[0]
 
-        title = metadata.at[text_id, 3].decode('utf-8')
-        author = metadata.at[text_id, 4]
-        # compensate for the fact that the author field is sometines empty
-        if author:
-            author = author.decode('utf-8')
-        source = u'{} - {}'.format(title, author)
+        out_file = os.path.join(output_dir, '{}.json'.format(text_id))
+        if not os.path.isfile(out_file):
+            title = metadata.at[text_id, 3].decode('utf-8')
+            author = metadata.at[text_id, 4]
+            # compensate for the fact that the author field is sometines empty
+            if author:
+                author = author.decode('utf-8')
+            source = u'{} - {}'.format(title, author)
 
-        print text_id
-        print source
+            print text_id
+            print source
 
-        if zipfile.is_zipfile(fi):
-            zf = zipfile.ZipFile(fi, 'r')
-            xml = zf.read(zf.namelist()[0])
-            soup = BeautifulSoup(xml, 'lxml')
+            if zipfile.is_zipfile(fi):
+                zf = zipfile.ZipFile(fi, 'r')
+                xml = zf.read(zf.namelist()[0])
+                soup = BeautifulSoup(xml, 'lxml')
+            else:
+                with open(fi, 'rb', encoding='utf-8') as f:
+                    soup = BeautifulSoup(f, 'lxml')
+
+            year = text2year[text_id]
+            num_sentences = get_num_sentences(soup)
+
+            events = process_emotions(soup, text_id, year, source,
+                                      emotion_labels, confidence)
+            json_out = {
+                'timeline': {
+                    'events': [],
+                    'sources': []
+                }
+            }
+
+            add_events(events, num_sentences, json_out)
+            add_source_text(soup, text_id, json_out)
+
+            # write output
+            with open(out_file, 'wb', encoding='utf-8') as f:
+                json.dump(json_out, f, sort_keys=True, indent=4)
+
+            end = time.time()
+            print 'processing took {} sec.'.format(end-start)
         else:
-            with open(fi, 'rb', encoding='utf-8') as f:
-                soup = BeautifulSoup(f, 'lxml')
-
-        year = text2year[text_id]
-        num_sentences = get_num_sentences(soup)
-
-        events, mention_counter = process_emotions(soup, text_id, year, source,
-                                                   emotion_labels, confidence)
-        num_events += len(mention_counter.keys())
-
-        end = time.time()
-        print 'processing took {} sec.'.format(end-start)
-
-        add_events(events, num_sentences, json_out)
-
-        print 'Now {} events in the JSON'.format(len(json_out['timeline']['events']))
-
-        add_source_text(soup, text_id, json_out)
-
-        # write intermediary output
-        temp_output_file = output_file.replace('.json', '{}.json'.format(i))
-        print temp_output_file
-        with open(temp_output_file, 'wb', encoding='utf-8') as f:
-            json.dump(json_out, f, sort_keys=True, indent=4)
-
-    # write output
-    with open(output_file, 'wb', encoding='utf-8') as f:
-        json.dump(json_out, f, sort_keys=True, indent=4)
+            print 'File already processed...'
 
 
 if __name__ == '__main__':
